Added manufactory which upgrade natural resources into fine products such as sugar cane -> molasses -> rum. Fine products can be sold for more.
Currently all manufactories are located on the colony, need to move some manufactories to the capital, ect...

Also need to add new pictures for new items.

Moved planet and crate UI into UI Pass which is rendered on top of everything. UI text should be drawn last so it appears on everything.